requirements.txt
asgiref==3.7.2
Django==4.2.3
django-autoslug==1.9.9
django-constance==2.9.1
django-picklefield==3.1
Pillow==10.0.0
sqlparse==0.4.4
